requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
pillow
psutil
httpx
sqlalchemy
aiosqlite
cairosvg
RPi.GPIO
spidev
